Split failed deletion tests
